backend/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
openai
numpy
pydantic
